update instance auth to permit admin key for all instances
