+++ cmd/persister-gw/main.go
@@ -2,12 +2,15 @@ package main
 
 import (
 	"flag"
+	"fmt"
 	"net/http"
 	"os"
+	"runtime"
 	"time"
 
 	"github.com/gorilla/handlers"
 	"github.com/gorilla/mux"
+	"github.com/grafana/globalconf"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/raintank/tsdb-gw/persister"
 	"github.com/raintank/tsdb-gw/util"
@@ -17,7 +20,11 @@ import (
 //Application: persister-gw
 
 var (
-	addr = flag.String("addr", "0.0.0.0:9001", "http service address")
+	app         = "persister-gw"
+	GitHash     = "(none)"
+	showVersion = flag.Bool("version", false, "print version string")
+	confFile    = flag.String("config", "/etc/gw/persister-gw.ini", "configuration file path")
+	addr        = flag.String("addr", "0.0.0.0:80", "http service address")
 )
 
 func indexHandler(w http.ResponseWriter, r *http.Request) {
@@ -28,6 +35,27 @@ func main() {
 	cfg := &persister.Config{}
 	util.RegisterFlags(cfg)
 	flag.Parse()
+	// Only try and parse the conf file if it exists
+	path := ""
+	if _, err := os.Stat(*confFile); err == nil {
+		path = *confFile
+	}
+
+	conf, err := globalconf.NewWithOptions(&globalconf.Options{
+		Filename:  path,
+		EnvPrefix: "GW_",
+	})
+	if err != nil {
+		log.Fatalf("error with configuration file: %s", err)
+		os.Exit(1)
+	}
+	conf.ParseAll()
+
+	if *showVersion {
+		fmt.Printf("%s (built with %s, git hash %s)\n", app, runtime.Version(), GitHash)
+		return
+	}
+
 	util.InitLogger()
 
 	p, err := persister.New(cfg)
