+++ scripts/config/persister-gw.ini
@@ -1,0 +1,10 @@
+addr="0.0.0.0:9001"
+bigtable-instance="persister"
+bigtable-prefix=
+bigtable-project="persister"
+bigtable-tablename="persister"
+gateway-addr="localhost:80/metrics"
+gateway-key="not_very_secret_key"
+log-level=2
+persister-buffer-size=1000
+persister-interval=60
