+++ scripts/deploy_docker.sh
@@ -8,6 +8,7 @@ cd ${DIR}
 VERSION=`git describe --always`
 
 docker push raintank/tsdb-gw:$VERSION
-# docker push raintank/tsdb-gw:latest
 
 docker push raintank/cortex-gw:$VERSION
+
+docker push raintank/persister-gw:$VERSION
